2461-maximum-sum-of-distinct-subarrays-with-length-k 08-23-2025

+++ LeetCode/Medium/2461-maximum-sum-of-distinct-subarrays-with-length-k/2461-maximum-sum-of-distinct-subarrays-with-length-k.cpp
@@ -5,7 +5,7 @@ public:
         unordered_set<int> seen;
         int left = 0;
         int right = 0;
-        int currSum = 0;
+        long long currSum = 0;
         while (right < nums.size()) {
             if (right - left + 1 < k) {
                 if (seen.find(nums[right]) == seen.end()) {
